% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_methods.R
\name{calc_subset_methods}
\alias{calc_subset_methods}
\title{Calculate the theoretical error for subsets of the pool of methods
Calculates the error for each possible subset of method in the processed dataset}
\usage{
calc_subset_methods(dataset)
}
\arguments{
\item{dataset}{A list of time series, with each element having at least
the entry \code{errors}, a vector with the errors produced by
the methods in the pool}
}
\value{
A list of list, the first dimension is the size of of subset
                        and the second the results of each subset,
                       having two entries: \code{iset} the indices of the subset in the
                       \code{errors} vector and \code{error}, the minimum error produced
                       by that subset.
}
\description{
Calculate the theoretical error for subsets of the pool of methods
Calculates the error for each possible subset of method in the processed dataset
}
