% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_classifier.R
\name{create_feat_classif_problem}
\alias{create_feat_classif_problem}
\title{Create a classification-like problem from a forecasting-processed time series dataset}
\usage{
create_feat_classif_problem(dataset)
}
\arguments{
\item{dataset}{A list with each element having a \code{features} and a optionally \code{errors} fields.
See \code{THA_feature} and \code{process_forecasts} for more information.}
}
\value{
\code{create_feat_classif_problem} returns a list with the entries:
\describe{
  \item{data}{The features extracted from the series.}
  \item{errors}{The errors produced by the forecasting method. If the air availabe in the dataset.}}
}
\description{
Create a classification-like problem from a forecasting-processed time series dataset
}
