% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forec_methods_list.R
\name{create_forec_method_list}
\alias{create_forec_method_list}
\alias{snaive_forec}
\alias{naive_forec}
\alias{auto_arima_forec}
\alias{ets_forec}
\alias{nnetar_forec}
\alias{tbats_forec}
\alias{stlm_ar_forec}
\alias{rw_drift_forec}
\title{Forecasting methods list
A list of the forecasting methods for use in the metalearnig process
The list follows the format described in the parameter \code{method_list}
of \code{\link{process_forecast_dataset}}}
\usage{
create_forec_method_list()

snaive_forec(x, h)

naive_forec(x, h)

auto_arima_forec(x, h)

ets_forec(x, h)

nnetar_forec(x, h)

tbats_forec(x, h)

stlm_ar_forec(x, h)

rw_drift_forec(x, h)
}
\arguments{
\item{x}{A \code{ts} object with the input time series}

\item{h}{The amount of future time steps to forecast}
}
\description{
Forecasting methods list
A list of the forecasting methods for use in the metalearnig process
The list follows the format described in the parameter \code{method_list}
of \code{\link{process_forecast_dataset}}
}
\section{Functions}{
\itemize{
\item \code{snaive_forec}: forecast::snaive

\item \code{naive_forec}: forecast::naive

\item \code{auto_arima_forec}: forecast::auto.arima

\item \code{ets_forec}: forecast::ets

\item \code{nnetar_forec}: forecast::nnetar

\item \code{tbats_forec}: forecast::tbats

\item \code{stlm_ar_forec}: forecast::stlm with ar modelfunction

\item \code{rw_drift_forec}: forecast::rwf
}}

\seealso{
\code{\link{process_forecast_dataset}}
}
