% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_classifier.R
\name{create_tempcv_dataset}
\alias{create_tempcv_dataset}
\title{Create Temporal Crossvalidated Dataset}
\usage{
create_tempcv_dataset(dataset)
}
\arguments{
\item{dataset}{A list with each element having at least the following
\describe{
  \item{x}{A time series object \code{ts} with the historical data.}
  \item{h}{The number of required forecasts.}
}}
}
\value{
A list with the same structure as the input,
 the following entries may be added or overwritten if existing
\describe{
  \item{x}{A time series object \code{ts} with the historical data.}
  \item{xx}{A time series with the true future data. Has length \code{h}
  unless the remaining \code{x} would be too short.}
}
}
\description{
Creates a datasets of time series for forecasting and metalearning
by extracting the final observations of each series
}
\details{
The final \code{h} observation are extrated an posed as the true future values
in the entry \code{xx} of the output list. At least 7 observations are kept as the
observable time series \code{x}.
}
