% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_classif_problem.R
\name{generate_THA_feature_dataset}
\alias{generate_THA_feature_dataset}
\title{Calculate features from Talagala, Hyndman, Athanaspoulos and add them to the dataset}
\usage{
generate_THA_feature_dataset(dataset, n.cores = 1)
}
\arguments{
\item{dataset}{A list the elements having a \code{ts} object with the name \code{x}}

\item{n.cores}{The number of cores to be used. \code{n.cores > 1} means parallel processing.}
}
\description{
For each series in \code{dataset}, the feature set used in
  (Talagala, Hyndman and Athanasopoulos, 2018) is calculated
   and added to the \code{forec_err_dataset}. A \code{tibble} named \code{THA_features}
    is added.
}
\examples{
processed <- generate_THA_feature_dataset(Mcomp::M3[c(1:3)], n.cores=1)
processed[[1]]$THA_features

}
