% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forec_methods_list.R
\name{M4_forec_methods}
\alias{M4_forec_methods}
\title{Forecasting methods list
A list of the forecasting methods for use in the metalearnig process
The list follows the format described in the parameter \code{methods}
of \code{{process_forecasts}}}
\usage{
M4_forec_methods()
}
\description{
Forecasting methods list
A list of the forecasting methods for use in the metalearnig process
The list follows the format described in the parameter \code{methods}
of \code{{process_forecasts}}
}
\seealso{
\code{\link{calc_forecasts}}
}
