% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combination_ensemble.R, R/ensemble_classifier.R
\name{metatemp_train}
\alias{metatemp_train}
\alias{train_selection_ensemble}
\alias{predict_selection_ensemble}
\alias{summary_performance}
\title{Train Metalearning Models
Trains a metalearning model by performing search on the hyperparameters}
\usage{
metatemp_train(train_dataset, eval_dataset, obj_fun = c("select",
  "combi:smax_abs", "combi:smax_sq", "combi:square"),
  filename = "meta_results.RData", verbose = FALSE, n.cores = 3)

train_selection_ensemble(data, errors, param = NULL)

predict_selection_ensemble(model, newdata)

summary_performance(predictions, dataset, print.summary = TRUE)
}
\arguments{
\item{train_dataset}{A list with elements in the metalearning format.
E.g. the output of a combination of \code{process_forecast_dataset} and \code{generate_THA_feature_dataset}
\describe{
  \item{x}{A time series object \code{ts} with the historical data.}
  \item{h}{The number of required forecasts.}
  \item{xx}{The number of required forecasts.}
  \item{THA_features}{The number of required forecasts.}
  \item{ff}{The number of required forecasts.}
  \item{errors}{The number of required forecasts.}
}}

\item{eval_dataset}{A list in the format of \code{train_dataset}, used for evaluating the error.
Can be the same as \code{train_dataset} to show training error}

\item{obj_fun}{The objective loss function that the metalearning minimizes}

\item{filename}{Name of the file used for saving the metalearning process}

\item{verbose}{Boolean indicating whether training progress messages may be printed}

\item{data}{A matrix with the input features data (extracted from the series).
One observation (the features from the original series) per row.}

\item{errors}{A matrix with the errors produced by each of the forecasting methods.
Each row is a vector with the errors of the forecasting methods.}

\item{model}{The xgboost model}

\item{newdata}{The feature matrix, one row per series}

\item{predictions}{A NXM matrix with N the number of observations(time series)
and M the number of methods.}

\item{dataset}{The list with the meta information, if given additional details will be provided}

\item{print.summary}{Boolean indicating wheter to print the information}

\item{labels}{A numeric vector from 0 to (nclass -1) with the targe labels for classification.}

\item{errors}{The NXM matrix with the erros per series and per method}

\item{labels}{Integer vector The true labels of the would be classification problem.
Possible values are from 0 to M-1}
}
\value{
\describe{
  \item{model}{The \code{xgboost} model found by the metalearning}
  \item{eval_log}{The log of hyperparametes tested with their produced errors}
}
}
\description{
Train Metalearning Models
Trains a metalearning model by performing search on the hyperparameters
}
\section{Functions}{
\itemize{
\item \code{train_selection_ensemble}: Train a method-selecting ensemble that minimizes forecasting error

\item \code{predict_selection_ensemble}: Produces predictions probabilities for the selection ensemble.

\item \code{summary_performance}: Analysis of the predictions
}}

