% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dataset.R
\name{process_forecast_dataset}
\alias{process_forecast_dataset}
\title{Generate Forecast and Errors for a time series dataset}
\usage{
process_forecast_dataset(dataset, methods_list, n.cores = 1)
}
\arguments{
\item{dataset}{The list containing the series. See details for the required format.}

\item{methods_list}{A list of strings with the names of the functions that generate
the forecasts.}

\item{n.cores}{The number of cores to be used. \code{n.cores > 1} means parallel processing.}
}
\value{
A list with the elements having the following structure
\describe{
  \item{x}{A time series object \code{ts} with the historical data.}
  \item{h}{The number of required forecasts.}
  \item{xx}{A time series of length \code{h} with the true future data.}
  \item{ff}{A matrix with F rows and \code{h} columns. Each row contains
  the forecasts of each method in \code{methods_list}}
  \item{errors}{A vector of F elements containing the OWI errors produced by each of the
  methods in \code{methods_list}}
  }
}
\description{
For each series in \code{dataset}, forecasts and
owi errors are generated for all methods in \code{methods_list}.
}
\details{
\code{dataset} must be a list with each element having the following format:
\describe{
  \item{x}{A time series object \code{ts} with the historical data.}
  \item{h}{The number of required forecasts.}
  \item{xx}{A time series of length code{h} with the true future data.}
}
\code{methods_list} is a list of strings with the names of the functions that generate the
forecasts. The functions must exist and take as parameters (\code{x}, \code{h}), with
\code{x} being the \code{ts} object with the input series and \code{h} the number of required
forecasts (after the last observation of \code{x}). The output of these functions must be
a vector or \code{ts} object of length \code{h} with the produced forecast.
No additional parameters are required in the functions.
}
\examples{
auto_arima_forec <- function(x, h) {
  model <- forecast::auto.arima(x, stepwise=FALSE, approximation=FALSE)
  forecast::forecast(model, h=h)$mean
}

snaive_forec <- function(x,h) {
  model <- forecast::snaive(x, h=length(x))
  forecast::forecast(model, h=h)$mean
}
rw_drift_forec <- function(x, h) {
  model <- forecast::rwf(x, drift=TRUE, h=length(x))
 forecast::forecast(model, h=h)$mean
}

create_example_list <- function() {
  methods_list <- list("auto_arima_forec")
  methods_list <- append(methods_list, "snaive_forec")
  methods_list <- append(methods_list, "rw_drift_forec")
  methods_list
}
methods <- create_example_list()
forec_results <- process_forecast_dataset(Mcomp::M3[1:4], methods, n.cores=1)

}
