% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_dataset.R
\name{process_forecasts}
\alias{process_forecasts}
\title{Generate Forecasts for a Time Series Dataset}
\usage{
process_forecasts(dataset, forecast_methods, chunk_size = 0,
  do_shuffle = TRUE, save_checkpoint_filename = NULL,
  load_checkpoint_filename = NULL)
}
\arguments{
\item{dataset}{The list containing the series. See details for the required format.}

\item{forecast_methods}{A list of strings with the names of the functions that generate
the forecasts.}
}
\value{
A list with the same structure as \code{dataset} but with the added entry
\describe{
  \item{ff}{A matrix with F rows and \code{h} columns. Each row contains
  the forecasts of each method in \code{methods} }
} to each element on a list
}
\description{
For each series in \code{dataset}, forecasts
are generated for all methods in \code{methods}.
}
\details{
\code{dataset} must be a list with each element having the following format:
\describe{
  \item{x}{A time series object \code{ts} with the historical data.}
  \item{h}{The number of required forecasts.}
}

\code{methods} is a list of strings with the names of the functions that generate the
forecasts. The functions must exist and take as parameters (\code{x}, \code{h}), with
\code{x} being the \code{ts} object with the input series and \code{h} the number of required
forecasts (after the last observation of \code{x}). The output of these functions must be
a vector or \code{ts} object of length \code{h} with the produced forecast.
No additional parameters are required in the functions.
}
\examples{
auto_arima_forec <- function(x, h) {
  model <- forecast::auto.arima(x, stepwise=FALSE, approximation=FALSE)
  forecast::forecast(model, h=h)$mean
}

snaive_forec <- function(x,h) {
  model <- forecast::snaive(x, h=length(x))
  forecast::forecast(model, h=h)$mean
}
rw_drift_forec <- function(x, h) {
  model <- forecast::rwf(x, drift=TRUE, h=length(x))
 forecast::forecast(model, h=h)$mean
}

create_example_list <- function() {
  methods <- list("auto_arima_forec")
  methods <- append(methods, "snaive_forec")
  methods <- append(methods, "rw_drift_forec")
  methods
}
methods <- create_example_list()
forec_results <- process_forecasts(Mcomp::M3[1:4], methods)

}
