% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{process_THA_features}
\alias{process_THA_features}
\alias{calc_features}
\title{Calculate features from Talagala, Hyndman, Athanaspoulos and add them to the dataset}
\usage{
process_THA_features(dataset, chunk_size = 0, do_shuffle = TRUE,
  save_checkpoint_filename = NULL, load_checkpoint_filename = NULL)

calc_features(seriesentry)
}
\arguments{
\item{dataset}{A list the elements having a \code{ts} object with the name \code{x}}

\item{chunk_size}{The size of the chunk. \code{chunk_size>0} will process \code{dataset} by chunks.}

\item{do_shuffle}{When processing by chunks, shuffle the data to get even comput. cost of the chunks}
}
\description{
For each series in \code{dataset}, the feature set used in
  (Talagala, Hyndman and Athanasopoulos, 2018) is calculated
   and added to the \code{forec_err_dataset}. A \code{tibble} named \code{features}
    is added.
}
\section{Functions}{
\itemize{
\item \code{calc_features}: Calculate features from Talagala, Hyndman, Athanaspoulos from a single time series
}}

\examples{
processed <- process_THA_features(Mcomp::M3[c(1:3)], n.cores=1)
processed[[1]]$features

}
