% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forec_methods_list.R
\name{rolling_forecast}
\alias{rolling_forecast}
\title{Make a rolling n-step forecast for a dataset in the format of the `FFORMA` format.}
\usage{
rolling_forecast(x, method, h = 4, n = 12)
}
\description{
Make a rolling n-step forecast for a dataset in the format of the `FFORMA` format.
}
