% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_classifier.R
\name{train_selection_ensemble}
\alias{train_selection_ensemble}
\title{Train a method-selecting ensemble that minimizes forecasting error}
\usage{
train_selection_ensemble(data, errors, param = NULL)
}
\arguments{
\item{data}{A matrix with the input features data (extracted from the series).
One observation (the features from the original series) per row.}

\item{errors}{A matrix with the errors produced by each of the forecasting methods.
Each row is a vector with the errors of the forecasting methods.}

\item{param}{speficic parameters to be passed to the xgboost::xgb.train function, otherwise a default value will be given}
}
\description{
Train a method-selecting ensemble that minimizes forecasting error
}
