% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_classifier.R
\name{train_selection_ensemble}
\alias{train_selection_ensemble}
\alias{predict_selection_ensemble}
\alias{summary_performance}
\title{Train a method-selecting ensemble that minimizes forecasting error}
\usage{
train_selection_ensemble(data, errors, labels)

predict_selection_ensemble(model, newdata)

summary_performance(predictions, errors, labels, dataset = NULL,
  print.summary = TRUE)
}
\arguments{
\item{data}{A matrix with the input features data (extracted from the series).
One observation (the features from the original series) per row.}

\item{errors}{A matrix with the errors produced by each of the forecasting methods.
Each row is a vector with the errors of the forecasting methods.}

\item{labels}{A numeric vector from 0 to (nclass -1) with the targe labels for classification.
ACTUALLY, THIS IS ALMOST IGNORED, MAY DISSAPEAR IN THE FUTURE.}
}
\description{
Train a method-selecting ensemble that minimizes forecasting error
}
\section{Functions}{
\itemize{
\item \code{predict_selection_ensemble}: Produces predictions probabilities for the selection ensemble.

\item \code{summary_performance}: Analysis of the predictions
}}

